chore: refactor statussqlfilter.go to respect views and have one file
