+++ pkg/cluster/statussqlfilter.go
@@ -9,6 +9,12 @@ import (
 	"github.com/kyma-incubator/reconciler/pkg/model"
 )
 
+const (
+	CreatedAtColumnName = "StatusCreatedAt"
+	StatusColumnName    = "Status"
+	RuntimeIDColumnName = "RuntimeID"
+)
+
 type statusSQLFilter interface {
 	Filter(dbType db.Type, colHdr *db.ColumnHandler) (string, error)
 }
@@ -18,7 +24,7 @@ type statusFilter struct {
 }
 
 func (sf *statusFilter) Filter(_ db.Type, statusColHdr *db.ColumnHandler) (string, error) {
-	statusColName, err := statusColHdr.ColumnName("Status")
+	statusColName, err := statusColHdr.ColumnName(CreatedAtColumnName)
 	if err != nil {
 		return "", err
 	}
@@ -38,11 +44,11 @@ type reconcileIntervalFilter struct {
 }
 
 func (rif *reconcileIntervalFilter) Filter(dbType db.Type, statusColHdr *db.ColumnHandler) (string, error) {
-	statusColName, err := statusColHdr.ColumnName("Status")
+	statusColName, err := statusColHdr.ColumnName(StatusColumnName)
 	if err != nil {
 		return "", err
 	}
-	createdColName, err := statusColHdr.ColumnName("Created")
+	createdColName, err := statusColHdr.ColumnName(CreatedAtColumnName)
 	if err != nil {
 		return "", err
 	}
@@ -53,6 +59,32 @@ func (rif *reconcileIntervalFilter) Filter(dbType db.Type, statusColHdr *db.Colu
 	case db.SQLite:
 		return fmt.Sprintf(`%s IN ('%s', '%s', '%s') AND %s <= DATETIME('now', '-%.0f SECONDS')`,
 			statusColName, model.ClusterStatusReady, model.ClusterStatusReconcileErrorRetryable, model.ClusterStatusDeleteErrorRetryable, createdColName, rif.reconcileInterval.Seconds()), nil
+	default:
+		return "", fmt.Errorf("database type '%s' is not supported by this filter", dbType)
+	}
+}
+
+type createdIntervalFilter struct {
+	runtimeID string
+	interval  time.Duration
+}
+
+func (rif *createdIntervalFilter) Filter(dbType db.Type, statusColHdr *db.ColumnHandler) (string, error) {
+	runtimeIDColName, err := statusColHdr.ColumnName(RuntimeIDColumnName)
+	if err != nil {
+		return "", err
+	}
+	createdColName, err := statusColHdr.ColumnName(CreatedAtColumnName)
+	if err != nil {
+		return "", err
+	}
+	switch dbType {
+	case db.Postgres:
+		return fmt.Sprintf(`%s = '%s' AND %s >= NOW() - INTERVAL '%.0f SECOND'`,
+			runtimeIDColName, rif.runtimeID, createdColName, rif.interval.Seconds()), nil
+	case db.SQLite:
+		return fmt.Sprintf(`%s = '%s' AND %s >= DATETIME('now', '-%.0f SECONDS')`,
+			runtimeIDColName, rif.runtimeID, createdColName, rif.interval.Seconds()), nil
 	default:
 		return "", fmt.Errorf("database type '%s' is not supported by this filter", dbType)
 	}
